@@ -11,46 +11,44 @@
     "clearCache": "rm -rf node_modules/.cache/ src/.temp .cache"
   },
   "dependencies": {
-    "@gridsome/plugin-sitemap": "^0.2.3",
-    "@sanity/client": "^0.144.0",
-    "axios": "^0.21.1",
+    "@gridsome/plugin-sitemap": "^0.4.0",
+    "@mapbox/rehype-prism": "^0.8.0",
+    "@sanity/client": "^3.0.5",
+    "axios": "^0.25.0",
     "cookieconsent": "^3.1.1",
-    "firebase-tools": "^9.10.2",
-    "gridsome": "^0.7.0",
-    "lodash": "^4.17.15",
-    "node-sass": "^4.13.0",
+    "firebase-tools": "^10.1.5",
+    "gridsome": "^0.7.23",
+    "lodash-es": "^4.17.21",
+    "node-sass": "^6",
     "pluralize": "^8.0.0",
     "pug": "^3.0.2",
-    "pug-plain-loader": "^1.0.0",
-    "rehype-autolink-headings": "^2.0.5",
-    "rehype-format": "^3.0.0",
-    "rehype-katex": "^2.0.0",
-    "rehype-prism": "^0.1.1",
-    "rehype-raw": "^4.0.1",
-    "rehype-stringify": "^6.0.1",
-    "remark-abbr": "^1.3.2",
-    "remark-attr": "^0.9.0",
-    "remark-containers": "^1.1.1",
-    "remark-emoji": "^2.0.2",
-    "remark-external-links": "^5.0.0",
-    "remark-fix-guillemets": "^1.0.20",
-    "remark-kbd": "^1.0.20",
-    "remark-parse": "^7.0.2",
-    "remark-rehype": "^5.0.0",
-    "remark-sectionize": "^1.1.0",
-    "remark-slug": "^5.1.2",
-    "sass-loader": "^8.0.0",
-    "typeface-fira-mono": "^0.0.72",
-    "typeface-montserrat": "^0.0.75",
-    "typeface-roboto": "^0.0.75",
-    "unified": "^8.4.2"
+    "pug-plain-loader": "^1.1.0",
+    "rehype-autolink-headings": "^6.1.1",
+    "rehype-format": "^4.0.1",
+    "rehype-katex": "^6.0.2",
+    "rehype-raw": "^6.1.1",
+    "rehype-stringify": "^9.0.3",
+    "remark-abbr": "^1.4.1",
+    "remark-directive": "^2.0.1",
+    "remark-emoji": "^3.0.2",
+    "remark-external-links": "^9.0.1",
+    "remark-fix-guillemets": "^1.1.1",
+    "remark-parse": "^10.0.1",
+    "remark-rehype": "^10.1.0",
+    "remark-sectionize": "^1.1.1",
+    "remark-slug": "^7.0.1",
+    "sass-loader": "^10",
+    "typeface-fira-mono": "^1.1.13",
+    "typeface-montserrat": "^1.1.13",
+    "typeface-roboto": "^1.1.13",
+    "unified": "^10"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.6.0",
-    "eslint-config-prettier": "^6.7.0",
-    "eslint-plugin-gridsome": "^1.4.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "prettier": "^1.19.1"
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.0.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-gridsome": "^1.6.2",
+    "eslint-plugin-prettier": "^4.0.0",
+    "prettier": "^2.5.1"
   }
 }
